+++ schedule/presentation/src/main/kotlin/vp/togedo/data/dto/fixedPersonalSchedule/FixedPersonalScheduleDto.kt
@@ -2,8 +2,8 @@ package vp.togedo.data.dto.fixedPersonalSchedule
 
 data class FixedPersonalScheduleDto(
     val id: String,
-    val startTime: Int,
-    val endTime: Int,
+    val startTime: Long,
+    val endTime: Long,
     val title: String,
     val color: String
 )
